refactor(frontend): extract fn to upload temporary file
