+++ apps/frontend/src/lib/utilities.ts
@@ -18,7 +18,7 @@ import {
 	IconMicrophone,
 } from "@tabler/icons-react";
 import { match } from "ts-pattern";
-import { gqlClient } from "./services/api";
+import { BASE_URL, gqlClient } from "./services/api";
 
 export const getSetColor = (l: SetLot) =>
 	match(l)
@@ -223,4 +223,27 @@ export const getPresignedGetUrl = async (key: string) => {
 		{ key },
 	);
 	return getPresignedS3Url;
+};
+
+export const uploadFileToServiceAndGetPath = async (
+	file: File,
+	onProgress: (event: ProgressEvent<XMLHttpRequestEventTarget>) => void,
+	onLoad: () => void,
+) => {
+	const formData = new FormData();
+	formData.append("files[]", file, file.name);
+	const data: string = await new Promise((resolve) => {
+		const xhr = new XMLHttpRequest();
+		xhr.upload.addEventListener("progress", (event) => {
+			if (event.lengthComputable) onProgress(event);
+		});
+		xhr.addEventListener("load", () => {
+			onLoad();
+			const data: string[] = JSON.parse(xhr.responseText);
+			resolve(data[0]);
+		});
+		xhr.open("POST", `${BASE_URL}/upload`, true);
+		xhr.send(formData);
+	});
+	return data;
 };
